backend/requirements.txt
fastapi
uvicorn
requests
pydantic
python-multipart
openai
google-generativeai
python-dotenv